# src/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CoCoPilot External Integration API
  description: |
    REST API for external tools to integrate with CoCoPilot.
    Provides flat (non-repo-nested) worker management, webhook
    registration for event callbacks, and system health monitoring.
  version: "1.0.0"
  contact:
    name: CoCoPilot Team
servers:
  - url: http://localhost:3000
    description: Local development daemon
paths:
...
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      summary: List registered webhooks
      tags: [Webhooks]
      responses:
        "200":
          description: Array of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
              example:
                - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  url: https://example.com/hooks/cocopilot
                  events: [worker.created, worker.completed]
                  createdAt: "2025-06-01T12:00:00.000Z"

    post:
      operationId: registerWebhook
      summary: Register a callback URL
      description: |
        Registers a webhook that will be called when the specified events
        occur. Webhooks are stored in-memory and reset on daemon restart.
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWebhookRequest"
            example:
              url: https://example.com/hooks/cocopilot
              events: [worker.created, worker.completed, pr.merged]
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Invalid request (bad URL, invalid events, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Invalid events: foo.bar. Valid events: worker.created, worker.updated, worker.completed, worker.failed, worker.removed, pr.created, pr.merged, ci.failed"
  /api/v1/webhooks/{id}:
    parameters:
      - $ref: "#/components/parameters/WebhookId"

    delete:
      operationId: removeWebhook
      summary: Remove a webhook
      tags: [Webhooks]
      responses:
        "204":
          description: Webhook removed
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Status
  # ---------------------------------------------------------------------------
...
components:
  parameters:
...
    WebhookId:
      name: id
      in: path
      required: true
      description: Webhook UUID
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    Webhook:
      type: object
      required: [id, url, events, createdAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - worker.created
              - worker.updated
              - worker.completed
              - worker.failed
              - worker.removed
              - pr.created
              - pr.merged
              - ci.failed
        createdAt:
          type: string
          format: date-time
    RegisterWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          description: HTTPS callback URL
        events:
          type: array
          minItems: 1
          items:
            type: string
            enum:
              - worker.created
              - worker.updated
              - worker.completed
              - worker.failed
              - worker.removed
              - pr.created
              - pr.merged
              - ci.failed
          description: Events to subscribe to
...